+++ activemodel/test/cases/secure_password_test.rb
@@ -290,6 +290,26 @@ class SecurePasswordTest < ActiveModel::TestCase
     assert_nil @user.password_salt
   end
 
+  test "password_cost" do
+    @user.password = "secret"
+    assert_equal @user.password_cost, BCrypt::Engine::MIN_COST
+  end
+
+  test "custom_cost_password_cost should return the configured custom cost" do
+    @user.custom_cost_password = "secret"
+    assert_equal @user.custom_cost_password_cost, BCrypt::Engine::MIN_COST + 1
+  end
+
+  test "password_cost should return nil when password is nil" do
+    @user.password = nil
+    assert_nil @user.password_cost
+  end
+
+  test "password_cost should return nil when password digest is nil" do
+    @user.password_digest = nil
+    assert_nil @user.password_cost
+  end
+
   test "Password digest cost defaults to bcrypt default cost when min_cost is false" do
     ActiveModel::SecurePassword.min_cost = false